HDDS-6938. move removeAcl test to TestOMPrefixAclRequest
